+++ src/gwt/src/org/rstudio/core/client/theme/res/ThemeStyles.java
@@ -50,6 +50,8 @@ public interface ThemeStyles extends CssResource
           
    String closeTabButton();
 
+   String fixedWidthFont();
+   
    String tabLayout();
    String tabLayoutLeft();
    String tabLayoutCenter();