+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTargetCppHelper.java
@@ -14,19 +14,38 @@
  */
 package org.rstudio.studio.client.workbench.views.source.editors.text;
 
+import org.rstudio.core.client.CommandWithArg;
+import org.rstudio.studio.client.RStudioGinjector;
+import org.rstudio.studio.client.common.GlobalDisplay;
+import org.rstudio.studio.client.common.GlobalProgressDelayer;
 import org.rstudio.studio.client.common.SimpleRequestCallback;
 import org.rstudio.studio.client.common.filetypes.TextFileType;
 import org.rstudio.studio.client.server.ServerError;
 import org.rstudio.studio.client.server.ServerRequestCallback;
-import org.rstudio.studio.client.workbench.views.buildtools.model.BuildServerOperations;
+import org.rstudio.studio.client.server.VoidServerRequestCallback;
 import org.rstudio.studio.client.workbench.views.source.editors.EditingTarget;
+import org.rstudio.studio.client.workbench.views.source.editors.text.ace.Position;
+import org.rstudio.studio.client.workbench.views.source.editors.text.cpp.CppCompletionContext;
 import org.rstudio.studio.client.workbench.views.source.model.CppCapabilities;
+import org.rstudio.studio.client.workbench.views.source.model.CppServerOperations;
+
+import com.google.inject.Inject;
 
 public class TextEditingTargetCppHelper
 {
-   public TextEditingTargetCppHelper(BuildServerOperations buildServer)
+   public TextEditingTargetCppHelper(CppCompletionContext completionContext,
+                                     DocDisplay docDisplay)
    {
-      buildServer_ = buildServer;
+      completionContext_ = completionContext;
+      docDisplay_ = docDisplay;
+      RStudioGinjector.INSTANCE.injectMembers(this);
+   }
+   
+   @Inject
+   void initialize(GlobalDisplay globalDisplay, CppServerOperations server)
+   {
+      globalDisplay_ = globalDisplay;
+      server_ = server;
    }
    
    public void checkBuildCppDependencies(
@@ -38,7 +57,7 @@ public class TextEditingTargetCppHelper
       if (!fileType.isC() || capabilities_.hasAllCapabiliites())
          return;
       
-      buildServer_.getCppCapabilities(
+      server_.getCppCapabilities(
                      new ServerRequestCallback<CppCapabilities>() {
          
          @Override
@@ -57,7 +76,7 @@ public class TextEditingTargetCppHelper
                      "are not currently installed");
                   
                   // do a prompted install of the build tools
-                  buildServer_.installBuildTools(
+                  server_.installBuildTools(
                            "Compiling C/C++ code for R",
                            new SimpleRequestCallback<Boolean>() {
                               @Override
@@ -90,7 +109,77 @@ public class TextEditingTargetCppHelper
    }
    
   
-   private final BuildServerOperations buildServer_;
+   public void findUsages()
+   {
+      cppCompletionOperation(new CppCompletionOperation() {
+         @Override
+         public void execute(String docPath, int line, int column)
+         {
+            server_.findCppUsages(
+                  docPath, 
+                  line, 
+                  column, 
+                  new CppCompletionServerRequestCallback(
+                                          "Finding usages..."));
+         }
+         
+      });
+   }
+   
+   private interface CppCompletionOperation
+   {
+      void execute(String docPath, int line, int column);
+   }
+   
+   private void cppCompletionOperation(final CppCompletionOperation operation)
+   {
+      if (completionContext_.isCompletionEnabled())
+      {
+         completionContext_.withUpdatedDoc(new CommandWithArg<String>() {
+            @Override
+            public void execute(String docPath)
+            {
+               Position pos = docDisplay_.getSelectionStart();
+               
+               operation.execute(docPath, 
+                                 pos.getRow() + 1, 
+                                 pos.getColumn() + 1);
+            }
+         });
+      }
+   }
+   
+   private class CppCompletionServerRequestCallback 
+                        extends VoidServerRequestCallback
+   {
+      public CppCompletionServerRequestCallback(String message)
+      {
+         super();
+         progressDelayer_ =  new GlobalProgressDelayer(
+               globalDisplay_, 1000, "Finding usages..");
+      }
+
+      @Override
+      public void onSuccess()
+      {
+         progressDelayer_.dismiss();
+      }
+
+      @Override
+      public void onFailure()
+      {
+         progressDelayer_.dismiss();
+      }
+
+      private final GlobalProgressDelayer progressDelayer_;
+   }
+
+   
+  
+   private GlobalDisplay globalDisplay_;
+   private CppServerOperations server_;
+   private final CppCompletionContext completionContext_;
+   private final DocDisplay docDisplay_;
    
    // cache the value statically -- once we get an affirmative response
    // we never check again